pyproject.toml
[[tool.poetry.source]]
name = "pypinlogic"
url = "http://pypi.nlogic.ai/simple/"

[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["MarkAntipin <antipinsuperstar@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.8"
python-dotenv = "^0.14.0"
python-multipart = "^0.0.5"
aiofiles = "^0.5.0"
tortoise-orm = "^0.16.16"
asyncpg = "^0.21.0"
nlogic = {version = "^4.9.5", extras = ["server", "image"]}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
